dev: 将文字层的选项移到UI层
