+++ lib/layers/uiLayer/assets/poli-light.png
