+++ lib/layers/uiLayer/utils.ts
@@ -1,0 +1,11 @@
+/**
+ * 按钮动画
+ * 
+*/
+function buttonAnimation(
+  elem: Element | Element[] | String,
+  scale: Number,
+  interval: Number
+) {}
+
+export { buttonAnimation };
