+++ src/App.vue
@@ -51,7 +51,7 @@ let showPlayer = ref(true)
 <template>
   <div style="display:flex;justify-content: center;">
     <div v-if="showPlayer">
-      <BaStoryPlayer :story="prologue" data-url="https://yuuka.cdn.diyigemt.com/image/ba-all-data" :width="width"
+      <BaStoryPlayer :story="prologue.splice(88)" data-url="https://yuuka.cdn.diyigemt.com/image/ba-all-data" :width="width"
         language="Cn" userName="testUser" :story-summary="storySummary" />
     </div>
     <div style="position: absolute;left: 0;display: flex;flex-direction: column;">
